Moved all methods from UserService which was using the DeliveryMan repository.
