+++ src/user/user.service.ts
@@ -8,7 +8,6 @@ import {
 } from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { User } from './entities/user.entity';
-import { DeliveryMan } from './entities/delivery-man.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CreateUserDto } from './dtos/user/create-user.dto';
 import { HashingService } from 'src/common/hashing/hashing.service';
@@ -31,8 +30,6 @@ export class UserService {
   constructor(
     @InjectRepository(User)
     private readonly userRepository: Repository<User>,
-    @InjectRepository(DeliveryMan)
-    private readonly deliveryManRepository: Repository<DeliveryMan>,
     private readonly hashingService: HashingService,
     private readonly roleService: RoleService,
     private readonly workTimeService: WorkTimeService,
@@ -232,25 +229,6 @@ export class UserService {
     return this.findOneByOrFail({ id: updated.id }, false);
   }
 
-  async updateMotoboyFields(
-    existsMotoboyData: boolean,
-    dto: UpdateUserDto,
-    user: User,
-  ) {
-    if (!existsMotoboyData) {
-      throw new BadRequestException('Dados do motoboy não enviados');
-    }
-
-    const motoboy = await this.findOneMotoboyByOrFail({ id: user.id });
-
-    // motoboy.motorcycle = dto.motorcycle ?? motoboy.motorcycle;
-    // MotorcycleService.update()
-
-    motoboy.daily = dto.daily ?? motoboy.daily;
-
-    return motoboy;
-  }
-
   async updatePassword(id: string, dto: UpdatePasswordDto) {
     const user = await this.findOneByOrFail({ id });
 
@@ -269,15 +247,6 @@ export class UserService {
     user.forceLogout = true;
 
     return this.saveUser(user);
-  }
-
-  async findAllMotoboy() {
-    const motoboys = await this.deliveryManRepository.find({
-      order: { createdAt: 'DESC' },
-      relations: essencial,
-    });
-
-    return motoboys;
   }
 
   async findAll({ role }: { role?: RoleEnum }) {
@@ -306,20 +275,6 @@ export class UserService {
       where: userData,
       relations: fields,
     });
-  }
-
-  async findOneMotoboyByOrFail(userData: Partial<User>, relations = true) {
-    const fields = relations ? full : essencial;
-    const motoboy = await this.deliveryManRepository.findOne({
-      where: userData,
-      relations: fields,
-    });
-
-    if (!motoboy) {
-      throw new NotFoundException('Motoboy não encontrado');
-    }
-
-    return motoboy;
   }
 
   findByEmail(email: string) {
@@ -357,21 +312,6 @@ export class UserService {
   async saveUser(user: Partial<User>) {
     const http400 = generateBadRequestException('Erro ao criar o usuário');
     const created = await this.userRepository
-      .save(user)
-      .catch((err: unknown) => {
-        if (err instanceof Error) {
-          this.logger.error(http400.message, err.stack);
-        }
-
-        throw http400;
-      });
-
-    return created;
-  }
-
-  async saveDeliveryMan(user: Partial<DeliveryMan>) {
-    const http400 = generateBadRequestException('Erro ao criar o motoboy');
-    const created = await this.deliveryManRepository
       .save(user)
       .catch((err: unknown) => {
         if (err instanceof Error) {
